requirements.txt
flask
flask_sqlalchemy
flask_migrate
python-dotenv
mysqlclient
pymysql